requirements.txt
fastapi
uvicorn
openai
faiss-cpu
python-dotenv
pydantic
langdetect
numpy
tqdm
python-docx
